DDF-3091 Extract ExpertimentalEnumerationExtractor
