+++ catalog/ui/catalog-ui-search/pom.xml
@@ -235,6 +235,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>ddf.ui</groupId>
+            <artifactId>catalog-ui-enumeration</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
